Added support for rtc_cmos (backported from branch/1.0-dev)
git-svn-id: http://svn.aircrack-ng.org/trunk@652 28c6078b-6c39-48e3-add9-af49d547ecab

+++ src/airdecap-ng.c
@@ -160,7 +160,7 @@ void calc_pmk( char *key, char *essid_pre, uchar pmk[40] )
 
     memset(essid,0,sizeof(essid));
     memcpy(essid,essid_pre,strlen(essid_pre));
-    slen = strlen(essid)+4;
+    slen = (int)strlen(essid)+4;
 
     /* setup the inner and outer contexts */
